+++ CHANGELOG.md
@@ -107,7 +107,7 @@ Main (unreleased)
 - Bump opentelemetry-collector and opentelemetry-collector-contrib versions from v0.80 to v0.85 (@wildum):
   - add `authoriy` attribute to `otelcol.exporter.loadbalancing` to override the default value in gRPC requests.
   - add `exemplars` support to `otelcol.connector.spanmetrics`.
-  - add `exclude_dimension` attribute to `otelcol.connector.spanmetrics` to exclude dimensions from the default set.
+  - add `exclude_dimensions` attribute to `otelcol.connector.spanmetrics` to exclude dimensions from the default set.
   - add `client_id_file` and `client_secret_file` attributes to `otelcol.auth.oauth2` to retrieve the ID and the secret via a file.
   - add `authority` attribute to `otelcol.receiver.otlp` to override the default value in gRPC requests.
   - add `disable_keep_alives` attribute to `otelcol.receiver.otlp` to disable the HTTP keep alive feature.